requirements.txt
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-google-genai
langchain-huggingface
sentence-transformers
faiss-cpu
python-dotenv
pypdf
docx2txt